requirements.txt
streamlit
pymupdf
langchain
langchain-community
openai
faiss-cpu
sentence-transformers
ollama